+++ plugins/auth-backend-module-pinniped-provider/src/index.ts
@@ -21,4 +21,5 @@
  */
 
 export { pinnipedAuthenticator, PinnipedStrategyCache } from './authenticator';
-export { authModulePinnipedProvider } from './module';
+export { authModulePinnipedProvider as default } from './module';
+export * from './deprecated';