+++ plugins/kubernetes-backend/src/auth/AzureIdentityStrategy.ts
@@ -15,12 +15,16 @@
  */
 
 import { Logger } from 'winston';
-import { AuthenticationStrategy, KubernetesCredential } from './types';
 import {
   AccessToken,
   DefaultAzureCredential,
   TokenCredential,
 } from '@azure/identity';
+import {
+  AuthMetadata,
+  AuthenticationStrategy,
+  KubernetesCredential,
+} from '@backstage/plugin-kubernetes-node';
 
 const aksScope = '6dae42f8-4368-4678-94ff-3960e28e3630/.default'; // This scope is the same for all Azure Managed Kubernetes
 
@@ -88,5 +92,9 @@ export class AzureIdentityStrategy implements AuthenticationStrategy {
 
   private tokenExpired(): boolean {
     return Date.now() >= this.accessToken.expiresOnTimestamp;
+  }
+
+  public presentAuthMetadata(_authMetadata: AuthMetadata): AuthMetadata {
+    return {};
   }
 }