+++ packages/create-app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/create-app
 
+## 0.5.9
+
+### Patch Changes
+
+- c9f71fb: Bumped create-app version.
+- ac277f3: Bumped create-app version.
+- 7acbb5a: Removed `mock-fs` dev dependency.
+- Updated dependencies
+  - @backstage/cli-common@0.1.13
+
+## 0.5.9-next.2
+
+### Patch Changes
+
+- Bumped create-app version.
+
 ## 0.5.9-next.1
 
 ### Patch Changes