+++ plugins/analytics-module-newrelic-browser/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-analytics-module-newrelic-browser
 
+## 0.0.6
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/core-components@0.13.10
+  - @backstage/core-plugin-api@1.8.2
+  - @backstage/config@1.1.1
+
 ## 0.0.6-next.1
 
 ### Patch Changes