+++ plugins/auth-backend-module-pinniped-provider/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @backstage/plugin-auth-backend-module-pinniped-provider
 
+## 0.1.3
+
+### Patch Changes
+
+- 928efbc: Deprecated the `authModulePinnipedProvider` export. A default export is now available and should be used like this in your backend: `backend.add(import('@backstage/plugin-auth-backend-module-pinniped-provider'));`
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.9
+  - @backstage/plugin-auth-node@0.4.3
+  - @backstage/config@1.1.1
+
+## 0.1.3-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.9-next.2
+  - @backstage/plugin-auth-node@0.4.3-next.2
+
 ## 0.1.3-next.1
 
 ### Patch Changes