+++ plugins/kubernetes-cluster/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @backstage/plugin-kubernetes-cluster
 
+## 0.0.5
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/core-components@0.13.10
+  - @backstage/core-plugin-api@1.8.2
+  - @backstage/plugin-kubernetes-react@0.2.1
+  - @backstage/plugin-kubernetes-common@0.7.3
+  - @backstage/plugin-catalog-react@1.9.3
+  - @backstage/catalog-model@1.4.3
+
+## 0.0.5-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.9.3-next.2
+
 ## 0.0.5-next.1
 
 ### Patch Changes