+++ plugins/shortcuts/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-shortcuts
 
+## 0.3.18
+
+### Patch Changes
+
+- 4016f21: Remove some unused dependencies
+- Updated dependencies
+  - @backstage/core-components@0.13.10
+  - @backstage/core-plugin-api@1.8.2
+  - @backstage/theme@0.5.0
+  - @backstage/types@1.1.1
+
 ## 0.3.18-next.1
 
 ### Patch Changes