+++ plugins/tech-insights-backend/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @backstage/plugin-tech-insights-backend
 
+## 0.5.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.20.1
+  - @backstage/catalog-client@1.5.2
+  - @backstage/plugin-tech-insights-node@0.4.15
+  - @backstage/backend-tasks@0.5.14
+  - @backstage/catalog-model@1.4.3
+  - @backstage/config@1.1.1
+  - @backstage/errors@1.2.3
+  - @backstage/types@1.1.1
+  - @backstage/plugin-tech-insights-common@0.2.12
+
+## 0.5.23-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.20.1-next.2
+  - @backstage/backend-tasks@0.5.14-next.2
+  - @backstage/plugin-tech-insights-node@0.4.15-next.2
+
 ## 0.5.23-next.1
 
 ### Patch Changes